Make it better

+++ db/post.go
@@ -2,6 +2,7 @@ package db
 
 import (
 	"fmt"
+	"strconv"
 	"time"
 
 	"github.com/gin-contrib/sessions"
@@ -126,17 +127,38 @@ func GetPosts(limit int) []Post {
 }
 
 func DeletePost(id string, c *gin.Context) {
-	session := sessions.Default(c)
 	mydb := Connect()
-	mysql := `DELETE FROM Post WHERE id = ?`
-	_, err := mydb.Exec(mysql, id)
-	if err != nil {
-		session.AddFlash("មាន​បញ្ហា​ក្នុង​ការលុបការផ្សាយ!", "error")
-		session.Save()
-		fmt.Println("Error deleting post:", err)
-		return
+	session := sessions.Default(c)
+	userRole, _ := c.Get("userRole")
+	if userRole != "Admin" {
+		post := GetPost(id)
+		author, _ := c.Get("userId")
+		if post.Author != author {
+			session.AddFlash("អ្នក​មិន​មាន​សិទ្ធិ​លុប​ការ​ផ្សាយ​របស់​អ្នក​ដទៃទេ!", "error")
+			session.Save()
+			return
+		} else {
+			mysql := `DELETE FROM Post WHERE id = ? AND author = ?`
+			_, err := mydb.Exec(mysql, id, author)
+			if err != nil {
+				session.AddFlash("មាន​បញ្ហា​ក្នុង​ការលុបការផ្សាយ!", "error")
+				session.Save()
+				fmt.Println("Error deleting post:", err)
+				return
+			}
+			defer mydb.Close()
+		}
+	} else if userRole == "Admin" {
+		mysql := `DELETE FROM Post WHERE id = ?`
+		_, err := mydb.Exec(mysql, id)
+		if err != nil {
+			session.AddFlash("មាន​បញ្ហា​ក្នុង​ការលុបការផ្សាយ!", "error")
+			session.Save()
+			fmt.Println("Error deleting post:", err)
+			return
+		}
+		defer mydb.Close()
 	}
-	defer mydb.Close()
 	session.AddFlash("​ការផ្សាយត្រូវ​​បានលុប​​ដោយ​ជោគជ័យ!", "success")
 	session.Save()
 }
@@ -153,4 +175,80 @@ func GetPost(id string) Post {
 	}
 	defer mydb.Close()
 	return *post
+}
+
+func UpdatePost(c *gin.Context) {
+	mydb := Connect()
+	defer mydb.Close()
+	session := sessions.Default(c)
+
+	title := c.PostForm("title")
+	content := c.PostForm("content")
+	categories := c.PostForm("categories")
+	thumb := c.PostForm("thumb")
+	date := c.PostForm("date")
+	videos := c.PostForm("videos")
+	postID := c.Param("id")
+	userRole, _ := c.Get("userRole")
+	mysql := `UPDATE Post SET title = ?, content = ?, categories = ?, thumb = ?, date = ?, videos = ? WHERE id = ?`
+
+	if userRole != "Admin" {
+		post := GetPost(postID)
+		author, _ := c.Get("userId")
+		if post.Author != author {
+			session.AddFlash("អ្នក​មិន​មាន​សិទ្ធិ​កែប្រែ​ការ​ផ្សាយ​របស់​អ្នក​ដទៃទេ!", "error")
+			session.Save()
+			return
+		}
+		_, err := mydb.Exec(mysql, title, content, categories, thumb, date, videos, postID)
+		if err != nil {
+			session.AddFlash("មាន​បញ្ហា​ក្នុង​ការកែប្រែ​ការផ្សាយ!", "error")
+			session.Save()
+			fmt.Println("Error updating post:", err)
+			return
+		}
+	} else if userRole == "Admin" {
+		_, err := mydb.Exec(mysql, title, content, categories, thumb, date, videos, postID)
+		if err != nil {
+			session.AddFlash("មាន​បញ្ហា​ក្នុង​ការកែប្រែ​ការផ្សាយ!", "error")
+			session.Save()
+			fmt.Println("Error updating post:", err)
+			return
+		}
+	}
+	session.AddFlash("ការផ្សាយ​ត្រូវ​បាន​កែប្រែ​ដោយ​ជោគជ័យ!", "success")
+	session.Save()
+}
+
+func PaginatePosts(c *gin.Context, dashboard int, query int) []Post {
+	mydb := Connect()
+	defer mydb.Close()
+	var page int
+	if query != 0 {
+		page = query
+	} else {
+		pageStr := c.Param("page")
+		page, _ = strconv.Atoi(pageStr)
+	}
+
+	limit := dashboard
+	offset := (page - 1) * dashboard
+	post := &Post{}
+	mysql := `SELECT * FROM Post ORDER BY date DESC LIMIT ? OFFSET ?`
+	rows, err := mydb.Query(mysql, limit, offset)
+	if err != nil {
+		fmt.Println("Error querying database:", err)
+		return nil
+	}
+	defer rows.Close()
+	var posts []Post
+	for rows.Next() {
+		err := rows.Scan(&post.ID, &post.Title, &post.Content, &post.Categories, &post.Thumb, &post.Date, &post.Videos, &post.Author, &post.CreatedAt, &post.UpdatedAt)
+		if err != nil {
+			fmt.Println("Error scanning row:", err)
+			continue
+		}
+		posts = append(posts, *post)
+	}
+	return posts
 }
